# kyros-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kyros Backend",
    "description": "Wellness (Phase 1) + Clinic (Phase 2)",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/users/guest": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create Guest",
        "description": "Idempotent find-or-create by X-Device-Id. Returns 201 on first creation, 200 on repeat.",
        "operationId": "create_guest_v1_users_guest_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": true,
            "description": "16\u201364 char alphanumeric+dash device identifier for guest auth.",
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9\\-]{16,64}$"
            }
          }
        ]
      }
    },
    "/v1/users/me": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get Me",
        "operationId": "get_me_v1_users_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": true,
            "description": "16\u201364 char alphanumeric+dash device identifier for guest auth.",
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9\\-]{16,64}$"
            }
          }
        ]
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Update Me",
        "operationId": "update_me_v1_users_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": true,
            "description": "16\u201364 char alphanumeric+dash device identifier for guest auth.",
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9\\-]{16,64}$"
            }
          }
        ]
      }
    },
...
    "/v1/wellness/lab-reports/": {
      "post": {
        "tags": [
          "Wellness"
        ],
        "summary": "Upload Lab Report",
        "operationId": "upload_lab_report_v1_wellness_lab_reports__post",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_lab_report_v1_wellness_lab_reports__post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabReportRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": true,
            "description": "16\u201364 char alphanumeric+dash device identifier for guest auth.",
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9\\-]{16,64}$"
            }
          }
        ]
      },
      "get": {
        "tags": [
          "Wellness"
        ],
        "summary": "List Lab Reports",
        "operationId": "list_lab_reports_v1_wellness_lab_reports__get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": true,
            "description": "16\u201364 char alphanumeric+dash device identifier for guest auth.",
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9\\-]{16,64}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LabReportRead"
                  },
                  "title": "Response List Lab Reports V1 Wellness Lab Reports  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/wellness/lab-reports/{report_id}": {
      "get": {
        "tags": [
          "Wellness"
        ],
        "summary": "Get Lab Report",
        "operationId": "get_lab_report_v1_wellness_lab_reports__report_id__get",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Report Id"
            }
          },
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": true,
            "description": "16\u201364 char alphanumeric+dash device identifier for guest auth.",
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9\\-]{16,64}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabReportRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Wellness"
        ],
        "summary": "Update Lab Report",
        "operationId": "update_lab_report_v1_wellness_lab_reports__report_id__patch",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Report Id"
            }
          },
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": true,
            "description": "16\u201364 char alphanumeric+dash device identifier for guest auth.",
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9\\-]{16,64}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabReportUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabReportRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Wellness"
        ],
        "summary": "Delete Lab Report",
        "description": "Soft delete: status \u2192 'deleted'. File is kept; a cleanup job sweeps later.",
        "operationId": "delete_lab_report_v1_wellness_lab_reports__report_id__delete",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Report Id"
            }
          },
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": true,
            "description": "16\u201364 char alphanumeric+dash device identifier for guest auth.",
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9\\-]{16,64}$"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/wellness/lab-reports/{report_id}/file": {
      "get": {
        "tags": [
          "Wellness"
        ],
        "summary": "Download Lab File",
        "operationId": "download_lab_file_v1_wellness_lab_reports__report_id__file_get",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Report Id"
            }
          },
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": true,
            "description": "16\u201364 char alphanumeric+dash device identifier for guest auth.",
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9\\-]{16,64}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_lab_report_v1_wellness_lab_reports__post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "metadata": {
            "type": "string",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "file",
          "metadata"
        ],
        "title": "Body_upload_lab_report_v1_wellness_lab_reports__post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LabReportRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "report_date": {
            "type": "string",
            "format": "date",
            "title": "Report Date"
          },
          "lab_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lab Name"
          },
          "file_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Url"
          },
          "file_mime": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Mime"
          },
          "parsed": {
            "items": {
              "$ref": "#/components/schemas/ParsedTest"
            },
            "type": "array",
            "title": "Parsed"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "source_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Ref"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "signed_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signed Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "report_date",
          "lab_name",
          "file_url",
          "file_mime",
          "parsed",
          "source",
          "source_ref",
          "note",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "LabReportRead"
      },
      "LabReportUpdate": {
        "properties": {
          "report_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Report Date"
          },
          "lab_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lab Name"
          },
          "parsed": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ParsedTest"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parsed"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "title": "LabReportUpdate"
      },
...
      "ParsedTest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "value": {
            "type": "string",
            "title": "Value"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "ref_low": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ref Low"
          },
          "ref_high": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ref High"
          },
          "flag": {
            "type": "string",
            "enum": [
              "normal",
              "low",
              "high",
              "critical"
            ],
            "title": "Flag"
          }
        },
        "type": "object",
        "required": [
          "name",
          "value",
          "unit",
          "flag"
        ],
        "title": "ParsedTest"
      },
...
      "UserRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "device_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "kyros_patient_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kyros Patient Id"
          },
          "subscription_tier": {
            "type": "string",
            "title": "Subscription Tier"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "device_id",
          "email",
          "kyros_patient_id",
          "subscription_tier",
          "timezone",
          "created_at",
          "updated_at"
        ],
        "title": "UserRead"
      },
      "UserUpdate": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          },
          "subscription_tier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subscription Tier"
          }
        },
        "type": "object",
        "title": "UserUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  },
...
}